requirements.txt
amqp==2.5.2
apiclient==1.0.4
apns2==0.7.1
billiard==3.6.3.0
cachetools==4.0.0
celery==4.4.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.9
Django==2.2.1
django-ckeditor==5.8.0
django-cleanup==4.0.0
django-cors-headers==3.2.0
django-github-s3==0.25
django-js-asset==1.2.2
django-mathfilters==0.4.0
django-push-notifications==2.0.0
django-redis==4.11.0
django-templated-mail==1.1.1
django-user-agents==0.4.0
djangorestframework==3.10.3
djoser==2.0.3
google-api-python-client==1.7.11
google-auth==1.11.0
google-auth-httplib2==0.0.3
h2==2.6.2
hpack==3.0.0
http-ece==1.1.0
httplib2==0.16.0
hyper==0.7.0
hyperframe==3.2.0
idna==2.9
importlib-metadata==1.6.0
kombu==4.6.8
oauth2client==4.1.3
Pillow==8.0.1
py-vapid==1.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pytz==2019.1
pywebpush==1.10.1
PyYAML==5.1.1
redis==3.4.1
requests==2.23.0
rsa==4.0
simplejson==3.16.0
six==1.14.0
sqlparse==0.3.0
ua-parser==0.8.0
uritemplate==3.0.1
urllib3==1.25.8
user-agents==2.0
vine==1.3.0
virtualenv==16.7.7
zipp==3.1.0
